+++ src/app/main/ConfigHandler.js
@@ -1,6 +1,8 @@
 import * as ConfigLoader from '../../config/ConfigLoader.js';
+import * as ConfigMaker from '../../app/client/menu/ConfigMaker.js';
 
 import * as ClientHandler from '../client/ClientHandler.js';
+import * as Menu from '../client/menu/Menu.js';
 
 /** If unable to load config file, null is returned. */
 export async function loadConfigFile(filepath)
@@ -32,4 +34,16 @@ export async function loadDefaultConfig(directory)
     };
     */
     return null;
+}
+
+export async function createNewConfig(directory)
+{
+    if (await Menu.askYesNo('Make a new config?'))
+    {
+        return ConfigMaker.run(directory, true);
+    }
+    else
+    {
+        return null;
+    }
 }