+++ src/test.js
@@ -1,0 +1,3 @@
+/** Run your tests here. */
+
+import './util/DateTest.js';